+++ mirsynth/src/main/groovy/io/github/mcolletta/mirsynth/DSP/Utils.groovy
@@ -31,14 +31,14 @@ class Utils {
 	
 	static float saw(double t, double T) {
 		double _t = t % T
-		return (_t / T) * 2.0f - 1.0f
+		return (float) ( (_t / T) * 2.0f - 1.0f )
 	}
 	
 	static float triangle(double t, double T) {
 		double _t = t % T
 		double T2 = T / 2
 		if (_t < T2)
-			return (_t / T2) * 2.0f - 1.0f
+			return (float) ( (_t / T2) * 2.0f - 1.0f )
 		else
 			return 1.0 - ( (_t - T2) / T2 * 2 - 1 )
 	}
@@ -68,11 +68,11 @@ class Utils {
 	}
 	
 	static float fm(double t, float A, float f0, float Ic, float fc) {
-		return A * sin( 2 * PI * f0 * t + (Ic * sin( 2 * PI * fc * t)) )
+		return (float) ( A * sin( 2 * PI * f0 * t + (Ic * sin( 2 * PI * fc * t)) ) )
 	}
 	
 	static float dfm(double t, float A, float I1, float I2, float f1, float f2) {
-		return A * sin( (I1 * sin(2 * PI * f1 * t)) + (I2 * sin(2 * PI * f2 * t)) )
+		return (float) ( A * sin( (I1 * sin(2 * PI * f1 * t)) + (I2 * sin(2 * PI * f2 * t)) ) )
 	}
 	
 	
@@ -80,7 +80,7 @@ class Utils {
 	//-----------------------------
 	
 	static float midiToFrequency(double d) {
-		return pow(2, (d - 6900) / 1200) * 440
+		return (float) ( pow(2, (d - 6900) / 1200) * 440 )
 	}
 	
 	static double millisToSamples(double millis, float sampleRate) {
